wallet: split itest CreateAccount invariants tests
